refactor(divide): change denominator type hint
John from business asks the denominator to be int

BREAKING CHANGE: need to change caller data type

+++ src/scratch_skatch.py
@@ -1,4 +1,4 @@
-def divide(numerator: float, denominator: float):
+def divide(numerator: float, denominator: int):
 	return numerator / denominator
